+++ app/src/main/java/hu/tb/minichefy/domain/model/recipe/entity/DatabaseEntities.kt
@@ -14,7 +14,7 @@ data class RecipeEntity(
     val icon: Int,
     val title: String,
     val quantity: Int,
-    val timeToCreate: Float,
+    val timeToCreate: Int,
     val timeUnit: TimeUnit,
     val ingredientList: FoodEntityWrapper
 )
